requirements.txt
# Core
python-dotenv==1.0.0
pydantic>=2.7.4
sqlalchemy==2.0.23
alembic==1.13.0

# Web & API
fastapi==0.108.0
uvicorn[standard]==0.25.0
requests==2.31.0
httpx==0.25.2

# Scraping & Data
googlemaps==4.10.0
pandas==2.1.4
beautifulsoup4==4.12.2

# AI & Automation
openai>=1.86.0

# Scheduling
APScheduler==3.10.4

# Database
aiosqlite==0.19.0

# Dashboard
streamlit==1.31.0
plotly==5.18.0

# Utils
pytz==2023.3